requirements.txt
Django==4.2
psycopg2-binary==2.9.6
django-environ==0.10.0
pre-commit==3.2.2
django-resized==1.0.2
pillow==9.5.0
social-auth-app-django==5.2.0
django-ckeditor==6.5.1
django-jazzmin==2.6.0
channels==4.0.0
daphne==4.0.0
django-filter==23.2
django-ratelimit==4.0.0
